FHIR-basics/requirements.txt
fhir.resources==6.1.0  # Versión sugerida para recursos FHIR
requests==2.31.0       # Para hacer solicitudes HTTP